pyproject.toml
[tool.poetry]
name = "llm-bridge"
version = "0.1.0"
description = "Streamlined access to multiple LLM providers through a unified interface and consistent response handling"
readme = "README.md"
authors = ["Yuval Merhav"]
license = "MIT"
packages = [{ include = "llm_bridge", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.78.1"
google-genai = "^1.15.0"
httpcore = "^1.0.9"
anthropic = "^0.51.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
pytest = "^8.1.1"
mypy = "^1.9.0"
pre-commit = "^3.6.2"

[build-system]
requires = ["setuptools", "wheel", "poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"
